fix(messaging): encode error events
